requirements.txt
pygame==2.5.2
gymnasium==0.29.1
torch==2.3.1
numpy==1.26.2
scipy==1.11.3
tensorboard==2.15.1
torch-geometric==2.4.0
qpsolvers==3.4.0
